requirements.txt
fastapi
uvicorn[standard]

sqlalchemy
pyodbc

passlib[bcrypt]
python-jose[cryptography]

pydantic
email-validator

python-dotenv

passlib[argon2]

langchain
langchain-core
langchain-community
pymupdf
sentence-transformers
chromadb
uuid